+++ src/biodwh2-aact/src/main/java/de/unibi/agbi/biodwh2/aact/model/BriefSummary.java
@@ -1,0 +1,14 @@
+package de.unibi.agbi.biodwh2.aact.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonPropertyOrder({"id", "nct_id", "description"})
+public class BriefSummary {
+    @JsonProperty("id")
+    public Long id;
+    @JsonProperty("nct_id")
+    public String nctId;
+    @JsonProperty("description")
+    public String description;
+}