+++ src/biodwh2-aact/src/main/java/de/unibi/agbi/biodwh2/aact/model/Link.java
@@ -1,0 +1,16 @@
+package de.unibi.agbi.biodwh2.aact.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonPropertyOrder({"id", "nct_id", "url", "description"})
+public class Link {
+    @JsonProperty("id")
+    public Long id;
+    @JsonProperty("nct_id")
+    public String nctId;
+    @JsonProperty("url")
+    public String url;
+    @JsonProperty("description")
+    public String description;
+}